pyproject.toml
[project]
name = "portfoliomaker"
version = "0.1.0"
description = ""
authors = [
    {name = "mallikarjun2823",email = "srinivasmalli15@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=6.0.4,<7.0.0)",
    "django-cors-headers (>=4.9.0,<5.0.0)",
    "pillow (>=12.2.0,<13.0.0)",
    "djangorestframework (>=3.17.1,<4.0.0)",
    "djangorestframework-simplejwt (>=5.4.0,<6.0.0)",
    "gunicorn (>=25.3.0,<26.0.0)",
    "whitenoise (>=6.12.0,<7.0.0)",
    "dj-database-url (>=3.1.2,<4.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "cloudinary (>=1.44.2,<2.0.0)",
    "django-cloudinary-storage (>=0.3.0,<0.4.0)",
    "requests (>=2.32.3,<3.0.0)",
    "pdfplumber (>=0.11.5,<0.12.0)",
    "python-docx (>=1.1.2,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
